+++ modules/item/sheets/item-sheet.js
@@ -1,7 +1,4 @@
-/**
- * Extend the basic ItemSheet with some very simple modifications
- * @extends {ItemSheet}
- */
+import { positionTooltip } from '../../sys/tooltip.js';
 
 export class HMItemSheet extends ItemSheet {
   /** @override */
@@ -46,6 +43,15 @@ export class HMItemSheet extends ItemSheet {
         super.activateListeners(html);
         if (!this.options.editable) return;
 
+        const root = html.get(0);
+        root.querySelectorAll('.tooltip').forEach((e) => {
+            e.addEventListener('mouseover', () => positionTooltip(e));
+        });
+        /*
+        root.querySelectorAll('.tooltip').forEach((e) => {
+            e.addEventListener('mouseover', () => positionTooltip(e));
+        });
+        */
         html.find('.editable').change(this._onEdit.bind(this));
 
         $(document).ready(() => $('.autoselect').focus(function autoselect() { $(this).select(); }));